requirements.txt
torch
torchvision
transformers==4.47.1
accelerate
peft==0.5.0
deepspeed==0.14.4
tiktoken
matplotlib
bitsandbytes
sentencepiece
wandb
protobuf
transformers_stream_generator
pillow
scipy
pandas
matplotlib
scikit-learn
opencv-python
tqdm
